Verfiy the PPA release result
This PR is a bug fix for release gpdb to PPA.
Although we can upload the gpdb deb package to the PPA,
it still maybe reject by the PPA due to some unexpected reasons.
Here, add a new job verify ppa release.
It can make the rest steps continue to run after releasing to PPA successfully

[#168673847]

Authored-by: Tingfang Bao <bbao@pivotal.io>