Only run postun when uninstall
postun is running for yum reinstall on rhel8, so that makes /usr/local/greenplum-db
removed after yum reinstall on rhel8. postun does not run for yum resintall on rhel6/7,
So inorder to fix the error, we make a if condition, only when $1=0 which means when
yum remove is called, the postun will be called.
Authored-by: Shaoqi Bai <bshaoqi@vmware.com>